requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
pytest==8.3.3
bcrypt==4.1.2
